Project_2/proj2_the_grading_mindset/requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
cachelib==0.9.0
click==8.1.3
colorama==0.4.5
Flask==2.2.0
Flask-Cors==3.0.10
Flask-Session==0.4.0
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg==3.0.16
psycopg-binary==3.0.16
psycopg-pool==3.1.1
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.8.2
python-dotenv==0.20.0
six==1.16.0
tomli==2.0.1
tzdata==2022.1
Werkzeug==2.2.1
zipp==3.8.1
